requirements.txt
tensorflow>=2.10.0
Pillow>=9.0.0
numpy>=1.21.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
opencv-python>=4.5.0
seaborn>=0.11.0
shap>=0.41.0  # Pour l'interprétabilité du modèle

# Development dependencies
pytest>=7.0.0
coverage>=6.0.0
pytest-cov>=3.0.0
black>=22.0.0  # For code formatting
flake8>=4.0.0  # For code linting
mypy>=0.900    # For type checking 